Do changes requested 1/2

+++ code/player/Player.Money.cs
@@ -5,14 +5,14 @@ namespace Cinema;
 public partial class Player
 {
     // The players current money
-    [Net] public double Money { get; set; }
+    [Net] public int Money { get; set; }
 
     /// <summary>
     /// Checks if a player can afford a certain amount of money.
     /// </summary>
     /// <param name="amount">How much money</param>
     /// <returns>True if money >= the amount</returns>
-    public bool CanAfford(double amount)
+    public bool CanAfford(int amount)
     {
         return Money >= amount;
     }
@@ -21,7 +21,7 @@ public partial class Player
     /// Adds money to the player's money.
     /// </summary>
     /// <param name="amount">The amount to add</param>
-    public void AddMoney(double amount)
+    public void AddMoney(int amount)
     {
         Money += amount;
     }
@@ -30,7 +30,7 @@ public partial class Player
     /// Takes money from the player.
     /// </summary>
     /// <param name="amount">How much to take</param>
-    public void TakeMoney(double amount)
+    public void TakeMoney(int amount)
     {
         Money -= amount;
 
@@ -41,8 +41,8 @@ public partial class Player
     /// <summary>
     /// Sets the amount of money the player has.
     /// </summary>
-    /// <param name="amount">How much money</param>
-    public void SetMoney(double amount)
+    /// <param name="amount">How much money to set</param>
+    public void SetMoney(int amount)
     {
         Money = amount;
     }
